Add pod lifecycle.
We were experiencing `agent lost` errors in our deployment pipelines
semi-regularly.
This is because when agent pods were asked to go down by our
cluster-autoscaler, they would stop immediately without allowing the
pipeline run to complete.

This commit adds a `preStop` lifecycle to the pod that sends a signal to
the agent to tell it to stop gracefully when it has finished running its
current job, and then waits for the agent to terminate.

This combined with increasing the `terminationGracePeriodSeconds` to a
value large enough to allow the pipelines to complete has eliminated
our broken builds due to the agents stopping prematurely.

The `terminationGracePeriodSeconds` is set to the Kubernetes default of
30 seconds, and would need to be increased by the end user for the
desired result (unless all their pipelines complete in < 30s that is).

Signed-off-by: Jim Barber <jim.barber@healthengine.com.au>