More configurability of fullname template variable
Provide more ways to tailor the `fullname` template variable that is
used for naming resources.
It's a way to independently set it from the `name` template variable
that is used in the `app` label and selector, which is useful when you
want to have multiple Buildkite agents with different configuration
deployed into the same Kubernetes namespace, but without having awkward
resource names.

See https://github.com/buildkite/charts/issues/99 for more details.